Updated Create Account Specifications
Made it so that First Name and Last Name cannot contain any special characters when creating a new account. Also made it so that the email that the user enters must be in a valid format. Also added unit tests for these conditions.

+++ Bug Tracker/Bug Tracker/Commands/ProjectsPage Commands/DeleteTicketCommand.cs
@@ -34,9 +34,12 @@ namespace Bug_Tracker.Commands.ProjectsPage_Commands
                 try
                 {
                     //clearing all comments in ticket before deleting ticket because cascade delete doesn't want to work
-                    foreach(Comment comment in ticketToDelete.Comments)
+                    if(ticketToDelete.Comments != null)
                     {
-                        await CommentDataService.Delete(comment.Id);
+                        foreach(Comment comment in ticketToDelete.Comments)
+                        {
+                            await CommentDataService.Delete(comment.Id);
+                        }
                     }
 
                     await TicketDataService.Delete(ticketToDelete.Id);
